Add file-vault-common
