+++ file-vault-config-source/pom.xml
@@ -18,13 +18,6 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>io.quarkus</groupId>
-        <artifactId>quarkus-bom</artifactId>
-        <version>${quarkus.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
         <groupId>io.quarkiverse.file-vault</groupId>
         <artifactId>quarkus-file-vault-config-source</artifactId>
         <version>${project.version}</version>
